+++ scripts/_TESTS/_tests-linear-ramp.js
@@ -18,6 +18,7 @@
 @import { getOrCreateCollection, setupModes, extractModes, processVariables } from "@Variables"
 @import { viewportLabel, namePrefix, resolveCollectionName, resolveGroup, readFoundation, writeManifest, readManifest, writeRegistry, normaliseConfig, toDomainConfig, foundationNamespace, foundationRegistryKey } from "@Foundation"
 @import { generateScale, isPiecewiseScaleType, snapScaleGrid } from "@Math Helpers"
+@import { scaleSequence, resolveModularRatio } from "@Scale Models"
 @import { spacingRampSpec, radiusRampSpec, ensureCompatRampConfig, materialiseRampTokens, materialiseRampSizes, validateRampScalingType, generateRampVariables, runLinearRamp, rampManifestSlice } from "@Linear Ramp"
 
 function currentRegistryRaw() {